feat(location): add location model
Add the location model.
The Location data class was defined, as well as the LocationRepository
interface defining the model for a location repository.
The Nominatim API was used for the LocationRepositoryNominatim
implementation.
To ensure that the app uses the same LocationRepository everywhere, a
LocationRepositoryProvider was added which provides a single repository.