fix(search): link home search bar to search screen
Link the home search bar to the search screen. This was done by
extracting the HomeSearchBar composable into a utility file to be shared
between the home screen and the search screen. The search bar on the
home screen is then just a button that navigates to the search screen.